DetectReduction: allow detection of array reductions on scf::ForOp
