+++ polygeist/lib/Dialect/Polygeist/Transforms/CMakeLists.txt
@@ -1,10 +1,10 @@
 add_mlir_dialect_library(MLIRPolygeistTransforms
   AffineCFG.cpp
-  AffineReduction.cpp
   ArgumentPromotion.cpp  
   BareMemRefToLLVM.cpp
   BarrierRemovalContinuation.cpp  
   CanonicalizeFor.cpp
+  DetectReduction.cpp
   InnerSerialization.cpp  
   LegalizeForSPIRV.cpp  
   LICM.cpp
